+++ src/test/java/roomescape/waiting/WaitingTest.java
@@ -1,4 +1,4 @@
-package roomescape.waiting.domain;
+package roomescape.waiting;
 
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import static org.assertj.core.api.AssertionsForClassTypes.assertThat;
@@ -11,6 +11,7 @@ import roomescape.exception.EscapeRoomException;
 import roomescape.reservationtime.domain.ReservationTime;
 import roomescape.slot.domain.Slot;
 import roomescape.theme.domain.Theme;
+import roomescape.waiting.domain.Waiting;
 
 public class WaitingTest {
